// backend/src/domain/enums/internship-status.enum.ts
export enum InternshipStatus {
  DRAFT = 'DRAFT',
  APPLIED = 'APPLIED',
  REVISION = 'REVISION',
  PENDING_EMPLOYER = 'PENDING_EMPLOYER',
  PENDING_COMMISSION = 'PENDING_COMMISSION',
  APPROVED = 'APPROVED',
  APPROVED_PENDING_SGK = 'APPROVED_PENDING_SGK',
  REJECTED = 'REJECTED',
  ONGOING = 'ONGOING',
  EVALUATION = 'EVALUATION',
  GRADED = 'GRADED',
  COMPLETED = 'COMPLETED',
  WITHDRAWN = 'WITHDRAWN',
  REVISION_REQUESTED = 'REVISION_REQUESTED',
}